+++ ricaun.Revit.Templates/tools/Templates/ItemTemplates/Revit/Controller/Controller.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Autodesk.Revit.DB;
+﻿using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
-using Autodesk.Revit.UI.Selection;
+using System;
 
 namespace $rootnamespace$
 {