+++ ricaun.Revit.Templates/tools/Templates/ItemTemplates/Revit/Mvvm/ViewModel/ViewModel.cs
@@ -1,0 +1,10 @@
+﻿using ricaun.Revit.Mvvm;
+using System;
+
+namespace $rootnamespace$
+{
+    public class $safeitemname$ : ObservableObject
+    {
+        
+    }
+}